flask-backend/requirements.txt
flask == 1.1.1
jinja2 == 2.10.1
google-cloud-vision == 0.41.0
google-cloud-translate == 2.0.0
google-cloud-storage == 1.24.1
Flask-RESTful
Flask-PyMongo